+++ src/Cool/Variables/Variable_vec3.h
@@ -17,7 +17,7 @@ inline auto imgui_widget(Variable<glm::vec3>& var) -> bool
 
 inline auto imgui_widget(VariableMetadata<glm::vec3>& meta) -> bool
 {
-    return ImGui::DragFloat("Drag speed", &meta.drag_speed, 0.01f, 0.001f, FLT_MAX / INT_MAX);
+    return ImGui::DragFloat("Drag speed", &meta.drag_speed, 0.01f, 0.001f, FLT_MAX / static_cast<float>(INT_MAX));
 }
 
 } // namespace Cool
